Refactor player container code
